Found bug in how NoteParser.characters handles the title. While xml.sax will stream contents, the title is being handled with an explicit setting each time.
This appears to cause problems when the .enex file has a note with a title containing an apostrophe (i.e., &pos;). The stream gets broken up into chunks of (1) before the apostrophe, (2) the apostrophe, and (3) after the apostrophe. As such, only the third chunk becomes the filename.

To fix this:
- note.py's Note class: Add append_title method, remove the set_title method, and move the call to create_filename into the finalize method.
- note_parser.py's NoteParser class: Change characters method to use Note.append_title instead of Note.set_title.

To verify this fix:
- The example Recipes.enex has notes whose titles include apostrophes.
- Add unittest test_run_parse_filename_match_note_title to fail when 1 exported recipe note file, whose title should include an apostrophe, does not exist.